Issue #37: support websockets
Implemented functions for WebSockets and Secure WebSockets support. The process
for treating HTTP requests in core.c has then been changed:

1. for every new request with a CONNECT,  create_http_socket() will peak into
the next request to determine if the next message contains binary (i.e
encrypted) content. If so, proceed with ssl/tls interception routine. This is
because WebSocket RFC forces the next message to be GET / blah if it is a
handshake.

2. for every request mark as `new` in core.c:proxenet_process_http_request(),
try to determine if the GET request is a WebSocket handshake. If so, update the
request headers (`proto_type` and `proto` fields) to mark the request as WS (or
WSS if ssl intercepted) so that further iteraction in
proxenet_process_http_request() main loop does not conflict with regular HTTP
(or HTTPS) requests.